pyproject.toml
[project]
name = "stage-three"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.120.3",
    "google-genai>=1.47.0",
]
